+++ src/main/java/gift/dto/ProductRegisterRequestDto.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.dto;
 
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
